+++ packages/meteor-type-validation/src/server/MeteorTypeValidation.ts
@@ -3,9 +3,6 @@ import { Meteor } from 'meteor/meteor';
 import { performance } from 'node:perf_hooks';
 import type Pino from 'pino';
 import { type GenericSchema, parse, ValiError } from 'valibot';
-import { formatValibotError } from './Errors';
-import type { GuardStatic } from './guards';
-import { Logger } from './Logger';
 import type {
     BaseContext,
     ContextWrapper,
@@ -18,7 +15,10 @@ import type {
     RateLimiterRule,
     ResourceType,
     WrappedContext,
-} from './ValidatedResources';
+} from './DefinitionTypeHelpers';
+import { formatValibotError } from './Errors';
+import type { GuardStatic } from './guards';
+import { Logger } from './Logger';
 
 export class MeteorTypeValidation<
     TAddedContext = {},
