+++ packages/meteor-type-validation/src/server/guards/Guard.ts
@@ -1,6 +1,6 @@
 import type { GenericSchema } from 'valibot';
 import * as v from 'valibot';
-import type { BaseContext, UnwrapSchemaOutput } from '../ValidatedResources';
+import type { BaseContext, UnwrapSchemaOutput } from '../DefinitionTypeHelpers';
 
 export abstract class Guard {
     constructor(
